+++ ReleaseNotes.md
@@ -6,6 +6,10 @@
 
 ol2: Click events didn't propagate properly when vector features were added from request to map. This has been fixed.
 
+### Admin layerselector
+
+SLD Style setup and management is added for wfs layers (versions 1.1.0 and 2.0.0) in admin layer selector.
+
 ## 1.36.1
 
 ### routingService