+++ ReleaseNotes.md
@@ -2,6 +2,10 @@
 
 ## 1.37.0
 
+### MarkersPlugin
+
+ol2 and ol3: Fixed own icon usable in request.
+
 ### VectorLayerPlugin
 
 ol2: Click events didn't propagate properly when vector features were added from request to map. This has been fixed.